# swagger.yaml
openapi: 3.0.3
info:
  title: Booking API
  description: API documentation for Booking Backend - Driver and Admin authentication, and booking management
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server (update with your actual URL)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for driver authentication
    bearerAuthAdmin:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
  schemas:
...
    Admin:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Admin User"
        email:
          type: string
          format: email
          example: "admin@example.com"
        role:
          type: string
          enum: [admin, super_admin]
          example: "admin"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Booking Schemas
...
    HotelCreate:
      type: object
      required:
        - name
        - address
        - zone
      properties:
        name:
          type: string
          minLength: 1
          example: "Grand Hotel Saigon"
        address:
          type: string
          minLength: 1
          example: "123 Nguyen Hue Street, District 1, Ho Chi Minh City"
        zone:
          type: string
          enum: [zone1, zone2, zone3]
          example: "zone1"
          description: "Zone for transport pricing calculation"
...
    Hotel:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Grand Hotel Saigon"
        address:
          type: string
          example: "123 Nguyen Hue Street, District 1, Ho Chi Minh City"
        zone:
          type: string
          enum: [zone1, zone2, zone3]
          example: "zone1"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Response Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
        errors:
          type: array
          items:
            type: string
          example: ["Validation error 1", "Validation error 2"]
paths:
...
  /api/admin/users:
    get:
      tags:
        - Admin User Management
      summary: Get all admin users
      description: Get all admin users with filters and pagination
      security:
        - bearerAuthAdmin: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
          example: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Number of items per page
          example: 20
        - in: query
          name: isActive
          schema:
            type: boolean
          description: Filter by active status
          example: true
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, super_admin]
          description: Filter by role
          example: "admin"
        - in: query
          name: search
          schema:
            type: string
          description: Search by name or email
          example: "john"
      responses:
        '200':
          description: Admin users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      admins:
                        type: array
                        items:
                          $ref: '#/components/schemas/Admin'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                            example: 50
                          page:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 20
                          pages:
                            type: integer
                            example: 3
        '401':
          description: Not authenticated or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Admin User Management
      summary: Create new admin user
      description: Create a new admin user account
      security:
        - bearerAuthAdmin: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                name:
                  type: string
                  example: "John Admin"
                  description: "Admin name (optional, defaults to email prefix)"
                email:
                  type: string
                  format: email
                  example: "admin@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "password123"
                role:
                  type: string
                  enum: [admin, super_admin]
                  default: admin
                  example: "admin"
      responses:
        '201':
          description: Admin user created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Admin user created successfully"
                  data:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/hotels:
    get:
      tags:
        - Admin Hotel Management
      summary: Get all hotels
      description: Get all hotels with pagination and filtering
      security:
        - bearerAuthAdmin: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search by name or address
        - in: query
          name: zone
          schema:
            type: string
            enum: [zone1, zone2, zone3]
          description: Filter by zone
        - in: query
          name: isActive
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: Hotels retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Hotel'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 50
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      totalPages:
                        type: integer
                        example: 5
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Admin Hotel Management
      summary: Create new hotel
      description: Create a new hotel record
      security:
        - bearerAuthAdmin: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelCreate'
      responses:
        '201':
          description: Hotel created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Hotel created successfully"
                  data:
                    $ref: '#/components/schemas/Hotel'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...